% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marathon_jerusalem}
\alias{marathon_jerusalem}
\title{The full set of runners for the Jerusalem marathon.}
\format{
A data frame with columns:
\describe{
 \item{age}{The variable is numeric. The age of the runner}
 \item{gender}{The variable is character. The gender of the runner (M/F)}
 \item{chiptime}{The variable is numeric. The time in minutes for the runner}
 \item{year}{The variable is numeric. The year of the marathon}
 \item{marathon}{The variable is character. The name of the marathon}
 \item{country}{The variable is character. The country where the marathon was held}
 \item{finishers}{The variable is numeric. The number of finishers at the marathon}
}
}
\source{
\url{http://faculty.chicagobooth.edu/george.wu/research/marathon/data.htm}
}
\usage{
marathon_jerusalem
}
\description{
This data set has ~2.5k observations.  Marathon website - https://jerusalem-marathon.com/en/home-page/
}
\examples{
\dontrun{
marathon_jerusalem
}
}
\keyword{datasets}
