% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marathon_nyc}
\alias{marathon_nyc}
\title{A random sample of 50% of males and females for each year of runners for all years of the New York City marathon where gender is recorded.}
\format{
A data frame with columns:
\describe{
 \item{age}{The variable is numeric. The age of the runner}
 \item{gender}{The variable is character. The gender of the runner (M/F)}
 \item{chiptime}{The variable is numeric. The time in minutes for the runner}
 \item{year}{The variable is numeric. The year of the marathon}
 \item{marathon}{The variable is character. The name of the marathon}
 \item{country}{The variable is character. The country where the marathon was held}
 \item{finishers}{The variable is numeric. The number of finishers at the marathon}
}
}
\source{
\url{http://faculty.chicagobooth.edu/george.wu/research/marathon/data.htm}
}
\usage{
marathon_nyc
}
\description{
This data set has just over 200k runners. The NYT had a good article - https://www.nytimes.com/2014/04/23/upshot/what-good-marathons-and-bad-investments-have-in-common.html?rref=upshot&_r=1. The NYC marathon website - https://www.nyrr.org/tcsnycmarathon
}
\examples{
\dontrun{
marathon_nyc
}
}
\keyword{datasets}
