% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_info}
\alias{race_info}
\title{Table of Information about Marathons}
\format{
A data frame with columns:
\describe{
 \item{year}{The variable is integer. The year of the marathon}
 \item{marathon}{The variable is character. The name of the marathon}
 \item{country}{The variable is character. The country where the marathon was held}
 \item{finishers}{The variable is integer. The number of finishers at the marathon}
 \item{mean_time}{The variable is numeric. The average finish time in minutes.}
}
}
\source{
\url{https://faculty.chicagobooth.edu/george.wu/research/marathon/marathon_names.htm}
}
\usage{
race_info
}
\description{
An interesting data set to see the effects of goals on what should be a unimodal distrubtion of finish times. The NYT had a good article - https://www.nytimes.com/2014/04/23/upshot/what-good-marathons-and-bad-investments-have-in-common.html?rref=upshot&_r=1
}
\examples{
\dontrun{
race_info
}
}
\keyword{datasets}
