% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_location}
\alias{race_location}
\title{}
\format{
A data frame with columns:
\describe{
 \item{marathon}{The variable is character. The name of the marathon that matches all other files}
 \item{marathon_name}{The variable is character. A cleaned name of the marathon}
 \item{state_id}{The variable is character. The two letter ID for each US state}
 \item{city}{The variable is character. The name of the city where the race is held}
 \item{finishers}{The variable is integer. The number of finishers at the marathon}
 \item{mean_time}{The variable is numeric. The average finish time in minutes.}
 \item{lat}{The variable is numeric. The lattitude of the city as listed at https://simplemaps.com/data/us-cities}
 \item{lng}{The variable is numeric. The longitude of the city as listed at https://simplemaps.com/data/us-cities}
 \item{elevation_m}{The variable is numeric. The elevation in meters above sea level as estimated from the elevatr R package.}
 \item{date}{The variable is Date. The approximate date of the marathon.  The year is correct but the month and day changes every year and we have marked it the same.}
 \item{month}{The variable is numeric. Approximate month of the marathon}
 \item{day}{The variable is numeric. Approximate day of the month of the marathon.}
 \item{year}{The variable is integer. The year of the marathon}
}
}
\source{
\url{http://faculty.chicagobooth.edu/george.wu/research/marathon/data.htm and https://simplemaps.com/data/us-cities}
}
\usage{
race_location
}
\description{
This data set has ~2k observations.
}
\examples{
\dontrun{
race_location
}
}
\keyword{datasets}
